Disable Google Analytics
Updated the site.google_analytics variable value to "false"; this should have been done a while ago when implemented for the main ryanbran.ch domain but the change was skipped over.

Also commented out the site.google_analytics_tracking_id variable entirely, to ensure that it is not accidentally implemented.